feat(remote): route Go runtime accept syscalls into remote-layer
Go issues syscalls directly from its own runtime rather than through
libc, so remote-layer's libc hooks never fire in a Go application: every
connection was accepted locally and the agent saw no incoming traffic.

Extract the Go syscall trampolines, runtime version detection and symbol
hooking out of mirrord-layer into a shared mirrord-layer-go crate, with
each layer plugging in its own dispatch table, and give remote-layer a
table routing accept, accept4, getsockname and getpeername into the same
detours the libc hooks use.

close and the dup family are still passed through, so claimed-socket
bookkeeping does not yet see fds a Go application closes.
