+++ mirrord/layer-go/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "mirrord-layer-go"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+publish.workspace = true
+edition.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+mirrord-layer-core = { path = "../layer/core" }
+
+libc.workspace = true
+nix = { workspace = true, features = ["signal"] }
+tracing.workspace = true
+
+[target.'cfg(target_os = "linux")'.dependencies]
+frida-gum.workspace = true
+syscalls.workspace = true
